Move Tuya converters
